rename to EditorData and extract logic
